% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tracked-time-issue.R
\name{add_tracked_time_issue}
\alias{add_tracked_time_issue}
\title{Tracked time to a issue}
\usage{
add_tracked_time_issue(base_url, api_key, owner, repo, id_issue, time)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The name of the repo}

\item{id_issue}{Index of the issue to add tracked time to}

\item{time}{The time to add the issue}
}
\description{
Add a tracked time to a issue
}
