% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-comment-issue.R
\name{create_comment_issue}
\alias{create_comment_issue}
\title{Adding a comment to an issue}
\usage{
create_comment_issue(base_url, api_key, owner, repo, id_issue, body)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}

\item{id_issue}{Index of the issue to create a comment}

\item{body}{The text that is added as a comment to the issue}
}
\description{
Add a comment to an issue in a gitea server
}
