% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-issue.R
\name{create_issue}
\alias{create_issue}
\title{Create an issue}
\usage{
create_issue(base_url, api_key, owner, repo, title, body)
}
\arguments{
\item{base_url}{URL prefix for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{Owner of the repo}

\item{repo}{Is the name of the repo}

\item{title}{Is the title of the issue}

\item{body}{Is the body of the issue}
}
\value{
list (invisibly) with the status result of the API
}
\description{
Create an issue
}
