% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-issue.R
\name{edit_issue}
\alias{edit_issue}
\title{Edit an issue}
\usage{
edit_issue(base_url, api_key, owner, repo, id_issue, body, state, title)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The name of the repo}

\item{id_issue}{Index of the issue to edit}

\item{body}{The time to add the issue}

\item{state}{The time to add the issue}

\item{title}{The time to add the issue}
}
\description{
Edit an issue
}
