% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-organizations.R
\name{get_admin_organizations}
\alias{get_admin_organizations}
\title{Returns organizations for an administrator user}
\usage{
get_admin_organizations(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Returns the list of organizations for a user with an
 administrator role
}
\details{
This function works only in the case that the `api_key` is
associated with a user with administrator role
}
\examples{
\dontrun{
get_admin_organizations(base_url = "https://example.gitea.service.com",
                        api_key = "b6026f861fd41a94c3389d54293de9d04bde6f7c")
}
}
