% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_an_issue.R
\name{get_an_issue}
\alias{get_an_issue}
\title{Returns an issue}
\usage{
get_an_issue(base_url, api_key, owner, repo, id_issue)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The reposository for the gitea service}

\item{id_issue}{Index of the issue to get}
}
\description{
Returns get an issue
}
