% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commits.R
\name{get_commits}
\alias{get_commits}
\title{Returns the repository commits}
\usage{
get_commits(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The name of the project where the repository exists}

\item{repo}{The repository name of which you want to obtain the information}
}
\description{
Returns the repository commits of the Gitea application
}
