% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-issues-closed-state.R
\name{get_issues_closed_state}
\alias{get_issues_closed_state}
\title{Returns the issues in closed state}
\usage{
get_issues_closed_state(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The user's owner for the gitea service}

\item{repo}{The reposository for the gitea service}
}
\description{
Returns the issues of the Gitea application
}
