% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_comments_issue.R
\name{get_list_comments_issue}
\alias{get_list_comments_issue}
\title{Returns get a list all comments on an issue}
\usage{
get_list_comments_issue(base_url, api_key, owner, repo, id_issue)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The name of the repo for the gitea service}

\item{id_issue}{Index of the issue to get}
}
\description{
Returns list all comments on an issue
}
