% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_comments_repository.R
\name{get_list_comments_repository}
\alias{get_list_comments_repository}
\title{Returns list of comments in a repository}
\usage{
get_list_comments_repository(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of repository for the gitea service}
}
\description{
Returns one list all comments in a repository
}
