% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list-an-org-teams.R
\name{get_list_org_teams}
\alias{get_list_org_teams}
\title{Returns organization's teams}
\usage{
get_list_org_teams(base_url, api_key, org)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}
}
\description{
Get organization's teams
}
\examples{
\dontrun{
get_list_org_teams(base_url = "https://example.gitea.service.com",
                   api_key = "3749s8SKFMA0kdi88jnaj8ska",
                   org = "company")
}
}
