% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list-users.R
\name{get_list_users}
\alias{get_list_users}
\title{Returns users of a gitea server}
\usage{
get_list_users(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
User list for a gitea server
}
\examples{
\dontrun{
get_list_users(base_url = "https://example.gitea.service.com",
               api_key = "3749s8SKFMA0kdi88jnaj8ska")
}
}
