% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-milestones.R
\name{get_milestones}
\alias{get_milestones}
\title{Returns open milestones of a repository}
\usage{
get_milestones(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}
}
\description{
Returns open milestones in a specific repository
}
\examples{
\dontrun{
get_milestones(base_url = "https://example.gitea.service.com",
          api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
          owner = "company",
          repo = "test_repo")
}
}
