% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-list-hooks.R
\name{get_org_list_hooks}
\alias{get_org_list_hooks}
\title{Return list an organization's webhooks}
\usage{
get_org_list_hooks(base_url, api_key, org)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}
}
\description{
List an organization's webhooks
}
