% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{Returns organizations of the user}
\usage{
get_organizations(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Returns list the current user's organizations
}
\examples{
\dontrun{
get_organizations(base_url = "https://example.gitea.service.com",
                  api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288")
}
}
