% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_version}
\alias{get_version}
\title{Returns the version}
\usage{
get_version(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Returns the version of the Gitea application
}
\examples{
get_version("https://try.gitea.io", "token 6ebcaefdaaf06aa7f59b4efc5faa4bcf1b56cfb1")
}
