% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitear.R
\docType{package}
\name{gitear}
\alias{gitear}
\title{\code{gitear} package}
\description{
'Gitea' is a community managed, lightweight code hosting solution 
 were projects and their respective git repositories can be managed 
 access and manage repositories, issues and organizations directly in R.
}
