% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tracked-time-issue.R
\name{add_tracked_time_issue}
\alias{add_tracked_time_issue}
\title{Add tracked time to an issue}
\usage{
add_tracked_time_issue(base_url, api_key, owner, repo, id_issue, time)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}

\item{id_issue}{Index of the issue to add tracked time to}

\item{time}{The time in seconds to add to the issue}
}
\description{
Add a tracked time to an issue
}
\examples{
\dontrun{
add_tracked_time_issue(base_url = "https://example.gitea.service.com",
                       api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
                       owner = "company",
                       repo = "test_repo",
                       id_issue = 2,
                       time = 15)
}
}
