% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-comment.R
\name{edit_comment}
\alias{edit_comment}
\title{Edit a comment}
\usage{
edit_comment(base_url, api_key, owner, repo, id_comment, body)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The name of the repo}

\item{id_comment}{Id of the comment to edit}

\item{body}{Is the content of comment}
}
\description{
Edit a comment
}
