% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-comment.R
\name{edit_comment}
\alias{edit_comment}
\title{Edit a comment}
\usage{
edit_comment(base_url, api_key, owner, repo, id_comment, body)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}

\item{id_comment}{Id of the comment to edit}

\item{body}{The text to replace the old comment}
}
\description{
Edit a comment in a specific issue
}
\examples{
\dontrun{
edit_comment(base_url = "https://example.gitea.service.com",
             api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
             owner = "company",
             repo = "test_repo",
             id_comment = 612,
             body = "This is the correction of my comment")
}
}
