% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-issue.R
\name{edit_issue}
\alias{edit_issue}
\title{Edit an issue}
\usage{
edit_issue(base_url, api_key, owner, repo, id_issue, title, body, state)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repo}

\item{repo}{The name of the repo}

\item{id_issue}{Index of the issue to edit}

\item{title}{The new issue title text}

\item{body}{The new issue body text}

\item{state}{The issue state}
}
\description{
Edit an issue
}
\examples{
\dontrun{
edit_issue(base_url = "https://example.gitea.service.com",
           api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
           owner = "company",
           repo = "test_repo",
           id_issue = 3,
           title = "My new title for this issue",
           body = "My new comment starts on this issue",
           state = "open")
}
}
