% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_an_organization.R
\name{get_an_organization}
\alias{get_an_organization}
\title{Return an organization}
\usage{
get_an_organization(org, base_url, api_key)
}
\arguments{
\item{org}{Name of the organization to get}

\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Get an organization
}
