% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-issues-closed-state.R
\name{get_issues_closed_state}
\alias{get_issues_closed_state}
\title{Returns the repository issues in closed state}
\usage{
get_issues_closed_state(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}
}
\description{
Returns the issues in closed state of a repository
}
\examples{
\dontrun{
get_issues_closed_state(base_url = "https://example.gitea.service.com",
                        api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
                        owner = "company",
                        repo = "test_repo")
}
}
