% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list-repos-of-an-organization.R
\name{get_list_repos_org}
\alias{get_list_repos_org}
\title{Organization's repositories}
\usage{
get_list_repos_org(base_url, api_key, org)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}
}
\description{
Get a list of an organization's repositories
}
