% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list-repos-of-an-organization.R
\name{get_list_repos_org}
\alias{get_list_repos_org}
\title{Returns organization's repositories}
\usage{
get_list_repos_org(base_url, api_key, org)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}
}
\description{
Get a list of an organization's repositories
}
\examples{
\dontrun{
get_list_repos_org(base_url = "https://example.gitea.service.com",
                    api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
                    org = "company")
}
}
