% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_repos_of_an_organization.R
\name{get_list_repos_org}
\alias{get_list_repos_org}
\title{Return list an organization's repos}
\usage{
get_list_repos_org(org, base_url, api_key)
}
\arguments{
\item{org}{Name of the organization to get repos}

\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Get list an organization's repos
}
