% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-hook.R
\name{get_org_hook}
\alias{get_org_hook}
\title{Returns a hook}
\usage{
get_org_hook(base_url, api_key, org, id_hook)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}

\item{id_hook}{Id of the hook to get information}
}
\description{
Get a hook information of a organizations
}
\examples{
\dontrun{
get_org_hook(base_url = "https://example.gitea.service.com",
             api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
             org = "company",
             id_hook = 2)
}
}
