% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_list_hooks.R
\name{get_org_list_hooks}
\alias{get_org_list_hooks}
\title{Return list an organization's webhooks}
\usage{
get_org_list_hooks(org, base_url, api_key)
}
\arguments{
\item{org}{Name of the organization}

\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
List an organization's webhooks
}
\examples{
get_org_list_hooks("Organizacion_1", "https://try.gitea.io", "6ebcaefdaaf06aa7f59b4efc5faa4bcf1b56cfb1")

}
