% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-list-hooks.R
\name{get_org_list_hooks}
\alias{get_org_list_hooks}
\title{Returns organization's webhooks}
\usage{
get_org_list_hooks(base_url, api_key, org)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{org}{Name of the organization}
}
\description{
Returns a list of organization's webhooks
}
\examples{
\dontrun{
get_org_list_hooks(base_url = "https://example.gitea.service.com",
             api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
             org = "company")
}
}
