% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{Returns the organizations of the user}
\usage{
get_organizations(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Returns list the current user's organizations
}
