% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-releases.R
\name{get_releases}
\alias{get_releases}
\title{Returns releases of a repository}
\usage{
get_releases(base_url, api_key, owner, repo)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}
}
\description{
Returns releasess in a specific repository
}
\examples{
\dontrun{
get_releases(base_url = "https://example.gitea.service.com",
             api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
             owner = "company",
             repo = "test_repo")
}
}
