% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{get_repositories}
\alias{get_repositories}
\title{Returns the repositories}
\usage{
get_repositories(base_url, api_key)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}
}
\description{
Returns the repositories of the Gitea application
}
\examples{
get_repositories("https://try.gitea.io", "token 6ebcaefdaaf06aa7f59b4efc5faa4bcf1b56cfb1")
}
