% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-times-issues.R
\name{get_times_issue}
\alias{get_times_issue}
\title{Returns issue's tracked times}
\usage{
get_times_issue(base_url, api_key, owner, repo, id_issue)
}
\arguments{
\item{base_url}{The base URL for your gitea server (no trailing '/')}

\item{api_key}{The user's API token key for the gitea service}

\item{owner}{The owner of the repository}

\item{repo}{The name of the repository}

\item{id_issue}{Index of the issue}
}
\description{
Returns a data frame of an issue's tracked times
}
\examples{
\dontrun{
get_times_issue(base_url = "https://example.gitea.service.com",
                api_key = "ccaf5c9a22e854856d0c5b1b96c81e851bafb288",
                owner = "company",
                repo = "test_repo",
                id_issue = 3)
}
}
